Move common functions to common.go
